+++ steelc/src/codegen/cleanup/scope.h
@@ -3,16 +3,18 @@
 #include <vector>
 #include <unordered_map>
 #include <string>
+#include <memory>
 
 #include <llvm/IR/BasicBlock.h>
-#include <llvm/IR/Instructions.h>
 
 #include "cleanup_action.h"
+#include "../memory/variable.h"
 
 struct scope {
 	llvm::BasicBlock* cleanup_block;
 	std::vector<cleanup_action> cleanup_actions;
+
 	bool cleanup_emitted = false;
 
-	std::unordered_map<std::string, llvm::AllocaInst*> locals_map;
+	std::unordered_map<std::string, std::shared_ptr<variable>> locals_map;
 };