+++ steelc/src/codegen/codegen_env.h
@@ -1,13 +1,23 @@
 #pragma once
 
 #include <vector>
+#include <string>
+#include <memory>
 
 #include <llvm/IR/IRBuilder.h>
 #include <llvm/IR/Module.h>
 #include <llvm/IR/Function.h>
 #include <llvm/IR/LLVMContext.h>
+#include <llvm/IR/BasicBlock.h>
+#include <llvm/IR/Instructions.h>
+#include <llvm/IR/Type.h>
+#include <llvm/IR/Value.h>
 
 #include "cleanup/scope.h"
+#include "cleanup/cleanup_action.h"
+#include "memory/lvalue.h"
+#include "memory/variable.h"
+#include "../parser/ast/expressions/expression.h"
 
 class codegen_env {
 public:
@@ -31,6 +41,9 @@ public:
 	// return value slot (for non-void functions)
 	llvm::AllocaInst* ret_slot = nullptr;
 
+	// create a new block in the current function
+	llvm::BasicBlock* make_block(const std::string& name, bool append_to_func = true);
+
 	// creates a new scope
 	void enter_scope();
 	// leaves the current scope, emitting cleanup code INLINE
@@ -42,12 +55,11 @@ public:
 	void alloc_return_slot(llvm::Type* ret_type);
 
 	// allocates a local variable
-	llvm::AllocaInst* alloc_local(llvm::Type* type, const std::string& name);
-	// registers a local variable onto the current scope's locals map (for lookup)
-	// NOTE: this is already called by alloc_local no need to recall
-	void register_local(const std::string& name, llvm::AllocaInst* local);
+	std::shared_ptr<variable> alloc_local(llvm::Type* type, const std::string& name);
 	// lookup a local variable by name in the current scope stack (from innermost to outermost)
-	llvm::AllocaInst* lookup_local(const std::string& name);
+	std::shared_ptr<variable> lookup_local(const std::string& name);
+
+	std::shared_ptr<lvalue> lvalue_from_expression(std::shared_ptr<expression> expr);
 
 	// emits a lifetime.start call
 	void emit_life_start(llvm::Value* ptr);
@@ -56,5 +68,6 @@ public:
 	void finalize_function();
 
 private:
+	llvm::AllocaInst* create_alloca(llvm::Type* type, const std::string& name = "");
 	void emit_cleanup(const cleanup_action& action);
 };