+++ steelc/src/codegen/codegen_visitor.h
@@ -13,6 +13,18 @@
 #include "builders/llvm_expression_builder.h"
 #include "codegen_env.h"
 #include "../parser/ast/ast_visitor.h"
+#include "../parser/ast/declarations/function_declaration.h"
+#include "../parser/ast/declarations/variable_declaration.h"
+#include "../parser/ast/expressions/assignment_expression.h"
+#include "../parser/ast/expressions/binary_expression.h"
+#include "../parser/ast/expressions/function_call.h"
+#include "../parser/ast/expressions/unary_expression.h"
+#include "../parser/ast/expressions/identifier_expression.h"
+#include "../parser/ast/expressions/literal.h"
+#include "../parser/ast/statements/block_statement.h"
+#include "../parser/ast/statements/control_flow/if_statement.h"
+#include "../parser/ast/statements/control_flow/for_loop.h"
+#include "../parser/ast/statements/control_flow/return_statement.h"
 
 class codegen_visitor : public ast_visitor {
 public:
@@ -35,11 +47,11 @@ public:
 	//void visit(std::shared_ptr<enum_option> option);
 	//void visit(std::shared_ptr<expression_statement> expr);
 	void visit(std::shared_ptr<binary_expression> expr);
-	//void visit(std::shared_ptr<assignment_expression> expr);
+	void visit(std::shared_ptr<assignment_expression> expr);
 	//void visit(std::shared_ptr<member_expression> expr);
 	//void visit(std::shared_ptr<address_of_expression> expr);
 	//void visit(std::shared_ptr<deref_expression> expr);
-	//void visit(std::shared_ptr<unary_expression> expr);
+	void visit(std::shared_ptr<unary_expression> expr);
 	//void visit(std::shared_ptr<index_expression> expr);
 	void visit(std::shared_ptr<identifier_expression> id) override;
 	//void visit(std::shared_ptr<this_expression> expr);
@@ -51,14 +63,11 @@ public:
 	void visit(std::shared_ptr<code_block> block) override;
 	void visit(std::shared_ptr<if_statement> if_stmt) override;
 	//void visit(std::shared_ptr<inline_if> inline_if);
-	//void visit(std::shared_ptr<for_loop> for_loop);
+	void visit(std::shared_ptr<for_loop> for_loop);
 	//void visit(std::shared_ptr<while_loop> while_loop);
 	void visit(std::shared_ptr<return_statement> ret_stmt) override;
 	//void visit(std::shared_ptr<break_statement> brk_stmt);
 	
-	
-	
-
 	llvm::Module& get_module() {
 		return module;
 	}
@@ -67,6 +76,8 @@ private:
 	llvm::LLVMContext context;
 	llvm::Module module;
 	llvm::IRBuilder<> builder;
+	// with the new accept method, NEVER directly use this field
+	// it should ONLY be set
 	llvm::Value* result;
 
 	llvm_type_converter type_converter;
@@ -76,4 +87,19 @@ private:
 	std::unique_ptr<codegen_env> env;
 
 	llvm::Value* generate_literal(std::shared_ptr<literal> lit);
+
+	bool op_requires_lvalue(token_type op);
+
+	// helper function to accept a node and return the resulting Value*
+	// this is much safer as it resets 'result' to nullptr before visiting
+	// preventing the accidental use of stale values
+	template<typename Nty>
+	llvm::Value* accept(std::shared_ptr<Nty>& node);
 };
+
+template<typename Nty>
+llvm::Value* codegen_visitor::accept(std::shared_ptr<Nty>& node) {
+	result = nullptr;
+	if (node) node->accept(*this);
+	return result;
+}